refactor : #188 Coupon 쿠폰 발행한 유저의 데이터베이스에 요청한 쿠폰 데이터 저장
- Quantity 타입을 Long으로 변경
- redis로 quantity 수량 조절이 끝나면 기존 데이터베이스의 quantity 업데이트
- 이후 TTL을 적용하지 않고, redis key를 삭제해서 메모리, 데이터 유실을 최소화

+++ com.taken_seat.coupon-service/src/main/java/com/taken_seat/coupon_service/application/dto/CouponDto.java
@@ -16,13 +16,13 @@ public class CouponDto {
 
     private String code;
 
-    private Integer quantity;
+    private Long quantity;
 
     private Integer discount;
 
     private LocalDateTime expiredAt;
 
-    public static CouponDto create(String name, String code, Integer quantity, Integer discount, LocalDateTime expiredAt) {
+    public static CouponDto create(String name, String code, Long quantity, Integer discount, LocalDateTime expiredAt) {
         return CouponDto.builder()
                 .name(name)
                 .code(code)
